+++ spec/lrama/grammar/lex_prec_spec.rb
@@ -3,49 +3,48 @@
 RSpec.describe Lrama::Grammar::LexPrec do
   let(:lex_prec) { Lrama::Grammar::LexPrec.new }
 
-  it "stores lex-prec rules" do
-    left = Lrama::Lexer::Token::Ident.new(s_value: "RANGLE")
-    right = Lrama::Lexer::Token::Ident.new(s_value: "RSHIFT")
+  def ident(name)
+    Lrama::Lexer::Token::Ident.new(s_value: name)
+  end
 
+  it "stores identity-right and longest-match rules" do
     lex_prec.add_rule(
-      left_token: left,
-      operator: Lrama::Grammar::LexPrec::SHORTER,
-      right_token: right,
+      left_token: ident("ID"),
+      operator: Lrama::Grammar::LexPrec::IDENTITY_RIGHT_LONGEST,
+      right_token: ident("IF"),
       lineno: 1
     )
 
     expect(lex_prec.rules.size).to eq(1)
-    expect(lex_prec.shorter_priority?("RANGLE", "RSHIFT")).to be true
-    expect(lex_prec.shorter_priority?("RSHIFT", "RANGLE")).to be false
+    expect(lex_prec.identity_precedes?("IF", "ID")).to be true
+    expect(lex_prec.identity_precedes?("ID", "IF")).to be false
+    expect(lex_prec.longest_pair?("ID", "IF")).to be true
   end
 
-  it "handles higher priority rules" do
-    left = Lrama::Lexer::Token::Ident.new(s_value: "IF")
-    right = Lrama::Lexer::Token::Ident.new(s_value: "ID")
-
+  it "does not infer transitive identity precedence" do
     lex_prec.add_rule(
-      left_token: left,
-      operator: Lrama::Grammar::LexPrec::HIGHER,
-      right_token: right,
+      left_token: ident("A"),
+      operator: Lrama::Grammar::LexPrec::IDENTITY_RIGHT,
+      right_token: ident("B"),
+      lineno: 1
+    )
+    lex_prec.add_rule(
+      left_token: ident("B"),
+      operator: Lrama::Grammar::LexPrec::IDENTITY_RIGHT,
+      right_token: ident("C"),
       lineno: 1
     )
 
-    expect(lex_prec.higher_priority?("IF", "ID")).to be true
-    expect(lex_prec.higher_priority?("ID", "IF")).to be false
+    expect(lex_prec.identity_precedes?("B", "A")).to be true
+    expect(lex_prec.identity_precedes?("C", "B")).to be true
+    expect(lex_prec.identity_precedes?("C", "A")).to be false
   end
 
-  it "handles same priority (lex-tie) rules" do
-    left = Lrama::Lexer::Token::Ident.new(s_value: "TOKEN_A")
-    right = Lrama::Lexer::Token::Ident.new(s_value: "TOKEN_B")
+  it "separates lexical ties from precedence" do
+    tie = Lrama::Grammar::LexTie.new
+    tie.add_tie("ID", "IF")
 
-    lex_prec.add_rule(
-      left_token: left,
-      operator: Lrama::Grammar::LexPrec::SAME_PRIORITY,
-      right_token: right,
-      lineno: 1
-    )
-
-    expect(lex_prec.same_priority?("TOKEN_A", "TOKEN_B")).to be true
-    expect(lex_prec.same_priority?("TOKEN_B", "TOKEN_A")).to be true
+    expect(tie.tied?("ID", "IF")).to be true
+    expect(lex_prec.identity_precedes?("IF", "ID")).to be false
   end
 end