+++ spec/lrama/pslr_family_regressions_spec.rb
@@ -55,7 +55,7 @@ RSpec.describe "PSLR family regressions" do
         %token-pattern X /x/
         %token-pattern IF /if/
         %token-pattern ID /[a-z]+/
-        %lex-prec IF - ID
+        %lex-prec ID <~ IF
 
         %%
 
@@ -186,7 +186,7 @@ RSpec.describe "PSLR family regressions" do
           %token-pattern ID /[a-z]+/
           %token-pattern RSHIFT />>/
           %token-pattern RANGLE />/
-          %lex-prec IF - ID
+          %lex-prec ID <~ IF
           %lex-prec RANGLE -s RSHIFT
 
           %%
@@ -240,7 +240,7 @@ RSpec.describe "PSLR family regressions" do
           %token-pattern ID /[a-z]+/
           %token-pattern RSHIFT />>/
           %token-pattern RANGLE />/
-          %lex-prec IF - ID
+          %lex-prec ID <~ IF
           %lex-prec RANGLE -s RSHIFT
 
           %%