+++ spec/lrama/states_spec.rb
@@ -3376,7 +3376,7 @@ RSpec.describe Lrama::States do
         %token-pattern X /x/
         %token-pattern IF /if/
         %token-pattern ID /[a-z]+/
-        %lex-prec IF - ID
+        %lex-prec ID <~ IF
 
         %%
 
@@ -3533,7 +3533,7 @@ RSpec.describe Lrama::States do
           %token-pattern ID /[a-z]+/
           %token-pattern RSHIFT />>/
           %token-pattern RANGLE />/
-          %lex-prec IF - ID
+          %lex-prec ID <~ IF
           %lex-prec RANGLE -s RSHIFT
 
           %%
@@ -3587,7 +3587,7 @@ RSpec.describe Lrama::States do
           %token-pattern ID /[a-z]+/
           %token-pattern RSHIFT />>/
           %token-pattern RANGLE />/
-          %lex-prec IF - ID
+          %lex-prec ID <~ IF
           %lex-prec RANGLE -s RSHIFT
 
           %%